Avoid duplicate submission of records
Currently in submitBatchWithRetry, when there were failure records, in
retry, the successful ones were submitted again. This caused firehose
sink got throttled and sometimes we got the "slow down" error code from
firehose delivery stream. In this case, the flush was considered as
"Failure" and then taking snapshot would be failed as well.